Move clickListener interfaces into contract
